feat: enhance jwt callback in auth module
The jwt callback in the auth module has been enhanced to not only log user and account information, but also to insert or update this data in the users and accounts database tables. This added functionality ensures that user and account data are consistently updated and maintained in our system.